Reorg table widgets into a new module
